+++ sf_core/src/config/mod.rs
@@ -9,6 +9,7 @@ pub mod logout;
 pub mod param_registry;
 pub mod param_store;
 mod private_key;
+mod token;
 pub use ini_loader::{IniConfig, get_ini_config, load_ini_files};
 pub use logging_config_loader::{logging_config_from_ini, logging_config_from_toml_section};
 pub use param_registry::ParamKey;